+++ src/main/java/james/parser/Parser.java
@@ -1,4 +1,18 @@
-package seedu;
+package james.parser;
+
+import james.JamesException;
+import james.command.addDeadlineCommand;
+import james.command.addEventCommand;
+import james.command.addListCommand;
+import james.command.addToDoCommand;
+import james.command.addMarkCommand;
+import james.command.addUnmarkCommand;
+import james.command.deleteCommand;
+import james.command.Command;
+import james.task.Deadline;
+import james.task.Event;
+import james.task.ToDo;
+
 
 import java.time.format.DateTimeParseException;
