+++ src/main/java/james/task/Deadline.java
@@ -1,4 +1,5 @@
-package seedu;
+package james.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
