+++ src/test/java/james/DeadlineTest.java
@@ -1,10 +1,10 @@
-package seedu;
+package james;
 
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class ToDoTest {
+public class DeadlineTest {
     @Test
     public void dummyTest(){
         assertEquals(2, 2);
