+++ src/test/java/james/ToDoTest.java
@@ -1,10 +1,10 @@
-package seedu;
+package james;
 
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class DeadlineTest {
+public class ToDoTest {
     @Test
     public void dummyTest(){
         assertEquals(2, 2);
